requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.23.5
tensorflow-cpu==2.11.0
nltk==3.8.1
scikit-learn==1.2.2
selenium==4.11.2
google-generativeai==0.4.1
gunicorn==21.2.0
beautifulsoup4==4.12.3
requests
